Modifications to package names and field definitions
