+++ lib/api/src/test/java/dev/aulait/qaa/api/RestrictedClient.java
@@ -1,0 +1,18 @@
+package dev.aulait.qaa.api;
+
+import lombok.AllArgsConstructor;
+
+@AllArgsConstructor
+public class RestrictedClient {
+
+  private RestClient client;
+  private String basePath;
+
+  public String get() {
+    return client.get(basePath, String.class);
+  }
+
+  public ErrorResponse getAsError() {
+    return client.get(basePath, ErrorResponse.class);
+  }
+}
